cena(pente): o roteiro da =49 diz ONDE o `Edge Flow` esta', e as quatro afirmacoes sobre a tela viram gate
Decisao do dono (2026-09-19), depois de ver o preco medido de mover a fileira:
«deixar a disposicao como esta' e escrever no roteiro onde o botao fica», com
uma etapa de arrumacao anunciada para depois.

MEDIDO (`diag_onde_cai_a_pista_do_pente`, encaixe 880 px):
  Tool 108 · Brush 534 · Radius 611 · **dobra 880** · Connected Only 922 ·
  Symmetry 1132 · Topology 1191 · ⭐ Edge Flow 1271 · Shading 1557 ·
  Scene 1790 · Bake 1945

⛔⛔ TRES afirmacoes minhas ao dono estavam erradas, e as tres eram sobre a
TELA: que a fileira vivia no `Pro` (e' **Basic**), que ela estava atra's do
sombreado (o sombreado esta' 286 px ABAIXO dela) e, implicita nessa, que a
ordem de pintura e' a da tabela `SECTIONS`. Ela NAO e': a `SECTIONS` declara
`Brush · Shading · Topology` e o `paint/body.rs` pinta
`… Symmetry → Topology → Shading …`. ⇒ *a ordem da tela le-se do `y`, nunca da
tabela* — a sonda passou a imprimi-la ORDENADA, com os sete cabecalhos.

O passo (1) deixou de dizer «na seccao Topology … logo abaixo esta' o Edge
Flow» (verdadeiro como ESTRUTURA e falso como GESTO) e passou a mandar rolar a
roda sobre o painel, nomeando o que se passa pelo caminho (`Tool`, `Brush`,
`Symmetry`) e o que fazer se a seccao estiver dobrada.

GATE `o_roteiro_da_49_diz_onde_o_edge_flow_esta` (na costura do painel, que e'
onde existe `paint`), quatro metades:
  (1) o NOME — os rotulos que o roteiro imprime sao os que o painel pinta;
  (2) o NIVEL — a fileira e' `Basic` com o carimbo na mao, logo o roteiro NAO
      manda trocar de nivel;
  (3) a DOBRA — escrita como EQUIVALENCIA (`y > 880` ⟺ o roteiro contem a
      frase da rolagem), ou seja uma CATRACA AO CONTRARIO: no dia em que a
      arrumacao trouxer a fileira para cima da dobra o gate REPROVA, e a cura
      e' apagar a frase, nunca afrouxar o numero;
  (4) a ORDEM — Tool → Brush → Symmetry → Topology → Edge Flow → Shading.

⇒ *uma cena que manda rolar a' procura de um controlo que ja' esta' a' vista
mente do mesmo jeito que uma que promete um efeito que nao acontece* (§5.0).

Mutacao 4 de 4, uma por metade (o rotulo do pente · o nivel da fileira · a
frase da rolagem · a ordem de pintura das seccoes), com controlo sobre o
proprio filtro e sobre a arvore restaurada.

PROVA: `nextest-impacted` 16 477/16 477 · a suite de costura do painel 75/75 ·
`cargo fmt` limpo · clippy `-D warnings` zero — com um aviso PRE-EXISTENTE
curado no caminho (`f64::from` sobre um `f64` na sonda de fabrica do
`ccda50553`): *um `-D warnings` que so' corre no fecho e' um portao que a wave
nao ve*.

Handoff §94 · CLAUDE.md §5.

+++ crates/ph2d-app-sculpt3d/src/scenes_pente_fabrica_tests.rs
@@ -71,7 +71,7 @@ fn diag_o_valor_de_fabrica_do_pente() {
             grade += g;
             f50 += fa;
             f90 += fb;
-            vinco += f64::from(v);
+            vinco += v;
             pior += pior_angulo(&m, &c, raio).0;
             lasca += lascas(&m, &c, raio, LIMIAR_DA_LASCA).0;
             relogio += dt;
